Split additional big cpp files
