+++ src/deviceconfig_validation.cpp
@@ -6,7 +6,9 @@
 //
 // This file is part of deviceconfig.cpp; see that file header for additional context.
 //
+// Split scope: DeviceConfig value validation routines.
 //---------------------------------------------------------------------------
+
 
 #include "globals.h"
