+++ src/effectmanager.cpp
@@ -142,89 +142,6 @@ void InitEffectsManager()
     }
 }
 
-void EffectManager::StartEffect()
-{
-    // Acquire both mutexes atomically. Separate sequential lock_guards form
-    // an AB-BA cycle with sites that use scoped_lock(render, effect_mgr).
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    // If there's a temporary effect override from the remote control active, we start that, else
-    // we start the current regular effect
-
-    if (!_tempEffect && _vEffects.empty())
-    {
-        _iCurrentEffect = 0;
-        _effectStartTime = millis();
-        debugV("No effect to start");
-        return;
-    }
-
-    if (!_vEffects.empty() && _iCurrentEffect >= _vEffects.size())
-        _iCurrentEffect = 0;
-
-    auto effect = _tempEffect ? _tempEffect : _vEffects[_iCurrentEffect];
-
-    #if USE_HUB75
-        auto& matrix = static_cast<HUB75GFX&>(*_gfx[0]);
-        matrix.SetCaption(effect->FriendlyName(), CAPTION_TIME);
-    #endif
-
-    // Zero the whites plane on every graphics device at effect-switch.
-    //
-    // CCT-aware effects (WarmGlowEffect and friends) write to whites[]
-    // every frame; effects that don't know about whites write only to
-    // leds[]. Without this reset, the previous effect's W-LED state
-    // persists through the transition and gets re-rendered by the
-    // PixelFormat under the new effect's colors - the strip ends up
-    // showing the old warm-white tint behind whatever colors the new
-    // effect draws. (leds[] doesn't need to be reset here for the same
-    // reason: every well-behaved effect overwrites the pixels it draws
-    // each frame, and effects that deliberately persist state across
-    // frames - like FireEffect's heat decay - want that persistence.
-    // The whites plane has no equivalent persisting-effect right now,
-    // so blanket-zeroing it on switch is the right semantic.)
-
-    for (auto& device : _gfx)
-    {
-        if (device && device->whites)
-            memset(device->whites, 0,
-                   device->GetLEDCount() * sizeof(CRGBW));
-    }
-
-    effect->Start();
-    _lastBeatSequence = g_Analyzer.LastBeat().sequence;
-    _lastNearBeatSequence = g_Analyzer.LastNearBeat().sequence;
-    _effectStartTime = millis();
-}
-
-void EffectManager::DispatchBeatIfNeeded()
-{
-#if ENABLE_AUDIO
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-
-    // Beat callbacks are sequenced here so every active effect sees the same
-    // detector output, including BeatEffectBase-derived effects via OnBeat().
-    if (!_tempEffect && _vEffects.empty())
-        return;
-
-    auto& currentEffect = GetCurrentEffect();
-
-    const auto nearBeat = g_Analyzer.LastNearBeat();
-    if (nearBeat.sequence != 0 && nearBeat.sequence != _lastNearBeatSequence)
-    {
-        currentEffect.OnNearBeat(nearBeat);
-        _lastNearBeatSequence = nearBeat.sequence;
-    }
-
-    const auto beat = g_Analyzer.LastBeat();
-    if (beat.sequence != 0 && beat.sequence != _lastBeatSequence)
-    {
-        currentEffect.OnBeat(beat);
-        _lastBeatSequence = beat.sequence;
-    }
-#endif
-}
-
 #ifndef NO_EFFECT_PERSISTENCE
     #define NO_EFFECT_PERSISTENCE 0
 #endif
@@ -294,274 +211,6 @@ EffectManager::~EffectManager()
     ClearEffects();
 }
 
-void EffectManager::SetTempEffect(std::shared_ptr<LEDStripEffect> effect)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-    _tempEffect = effect;
-}
-
-std::vector<std::shared_ptr<GFXBase>> & EffectManager::GetBaseGraphics()
-{
-    return _gfx;
-}
-
-void EffectManager::ReportNewFrameAvailable()
-{
-    // Listener vectors store references whose owners may deregister on other
-    // tasks. Hold this mutex while iterating so remove/destruction cannot race
-    // a callback dispatch.
-
-    std::lock_guard listenerGuard(_listenerMutex);
-    INFORM_EVENT_LISTENERS(_frameEventListeners, IFrameEventListener::OnNewFrameAvailable);
-}
-
-void EffectManager::AddFrameEventListener(IFrameEventListener& listener)
-{
-    std::lock_guard listenerGuard(_listenerMutex);
-    _frameEventListeners.emplace_back(listener);
-}
-
-void EffectManager::RemoveFrameEventListener(IFrameEventListener& listener)
-{
-    // Match by address — reference_wrapper has no operator== so we compare
-    // the underlying object pointers. erase/remove_if so we drop every entry
-    // even if the same listener was registered more than once.
-    // Robert may have a better or less obtuse way to do this, but it works
-    // and it's not like we have thousands of listeners.  Change at will!
-
-    std::lock_guard listenerGuard(_listenerMutex);
-    _frameEventListeners.erase(
-        std::remove_if(_frameEventListeners.begin(), _frameEventListeners.end(),
-                       [&](const std::reference_wrapper<IFrameEventListener>& w)
-                       { return &w.get() == &listener; }),
-        _frameEventListeners.end());
-}
-
-void EffectManager::AddEffectEventListener(IEffectEventListener& listener)
-{
-    std::lock_guard listenerGuard(_listenerMutex);
-    _effectEventListeners.emplace_back(listener);
-}
-
-// Must provide at least one drawing instance, like the first matrix or strip we are drawing on
-GFXBase& EffectManager::g(int iChannel)
-{
-    return *_gfx[iChannel];
-}
-
-const GFXBase& EffectManager::g(int iChannel) const
-{
-    return *_gfx[iChannel];
-}
-
-bool EffectManager::IsEffectEnabled(size_t i) const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    if (i >= _vEffects.size())
-    {
-        debugW("Invalid index for IsEffectEnabled");
-        return false;
-    }
-    return _vEffects[i]->IsEnabled();
-}
-
-void EffectManager::PlayAll(bool bPlayAll)
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    _bPlayAll = bPlayAll;
-}
-
-void EffectManager::SetInterval(uint interval, bool skipSave)
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-
-    // Reject/ignore intervals smaller than a second, but allow 0 (infinity)
-    if (interval > 0 && interval < 1000)
-        return;
-
-    _effectInterval = interval;
-
-    if (!skipSave)
-        SaveEffectManagerConfig();
-
-    {
-        std::lock_guard listenerGuard(_listenerMutex);
-        INFORM_EVENT_LISTENERS(_effectEventListeners, IEffectEventListener::OnIntervalChanged, interval);
-    }
-}
-
-std::vector<std::shared_ptr<LEDStripEffect>> EffectManager::EffectsList() const
-{
-    // Return a stable snapshot instead of a reference to the live vector. Web,
-    // debug, and render tasks can otherwise hold an iterator/reference while
-    // another task reorders or deletes effects.
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return _vEffects;
-}
-
-std::shared_ptr<LEDStripEffect> EffectManager::EffectAt(size_t index) const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return index < _vEffects.size() ? _vEffects[index] : nullptr;
-}
-
-bool EffectManager::IsCoreEffect(size_t index) const
-{
-    auto effect = EffectAt(index);
-    return effect && effect->IsCoreEffect();
-}
-
-size_t EffectManager::EffectCount() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return _vEffects.size();
-}
-
-bool EffectManager::AreEffectsEnabled() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return std::any_of(_vEffects.begin(), _vEffects.end(), [](const auto& pEffect){ return pEffect->IsEnabled(); } );
-}
-
-bool EffectManager::HasCurrentEffect() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return _tempEffect || !_vEffects.empty();
-}
-
-size_t EffectManager::GetCurrentEffectIndex() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return _iCurrentEffect;
-}
-
-LEDStripEffect& EffectManager::GetCurrentEffect() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    if (!_tempEffect && _vEffects.empty())
-        throw std::runtime_error("No current effect");
-
-    const size_t index = _vEffects.empty() ? 0 : std::min(_iCurrentEffect, _vEffects.size() - 1);
-    return *(_tempEffect ? _tempEffect : _vEffects[index]);
-}
-
-String EffectManager::GetCurrentEffectName() const
-{
-    // Return a copy while holding the effect lock. Returning a reference here
-    // was unsafe because another task can switch/delete/reinitialize effects
-    // immediately after the lock is released.
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    if (_tempEffect)
-        return _tempEffect->FriendlyName();
-
-    if (_vEffects.empty())
-        return "No Effect";
-
-    return _vEffects[std::min(_iCurrentEffect, _vEffects.size() - 1)]->FriendlyName();
-}
-
-void EffectManager::SetCurrentEffectIndex(size_t i)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (i >= _vEffects.size())
-    {
-        debugW("Invalid index for SetCurrentEffectIndex");
-        return;
-    }
-    _iCurrentEffect = i;
-    _effectStartTime = millis();
-
-    StartEffect();
-    SaveCurrentEffectIndex();
-
-    {
-        std::lock_guard listenerGuard(_listenerMutex);
-        INFORM_EVENT_LISTENERS(_effectEventListeners, IEffectEventListener::OnCurrentEffectChanged, i);
-    }
-}
-
-uint EffectManager::GetTimeUsedByCurrentEffect() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return millis() - _effectStartTime;
-}
-
-uint EffectManager::GetTimeRemainingForCurrentEffect() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    if (!_tempEffect && _vEffects.empty())
-        return 0;
-
-    // If the Interval is set to zero, we treat that as an infinite interval and don't even look at the time used so far
-    uint timeUsedByCurrentEffect = GetTimeUsedByCurrentEffect();
-    uint interval = GetEffectiveInterval();
-
-    return timeUsedByCurrentEffect > interval ? 0 : (interval - timeUsedByCurrentEffect);
-}
-
-uint EffectManager::GetEffectiveInterval() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    if (!_tempEffect && _vEffects.empty())
-        return IsIntervalEternal() ? std::numeric_limits<uint>::max() : _effectInterval;
-
-    const size_t index = _vEffects.empty() ? 0 : std::min(_iCurrentEffect, _vEffects.size() - 1);
-    auto& currentEffect = *(_tempEffect ? _tempEffect : _vEffects[index]);
-    // This allows you to return a MaximumEffectTime and your effect won't be shown longer than that
-    return min((IsIntervalEternal() ? std::numeric_limits<uint>::max() : _effectInterval),
-               (currentEffect.HasMaximumEffectTime() ? currentEffect.MaximumEffectTime() : std::numeric_limits<uint>::max()));
-}
-
-uint EffectManager::GetInterval() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return _effectInterval;
-}
-
-bool EffectManager::IsIntervalEternal() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return _effectInterval == 0;
-}
-
-void EffectManager::NextPalette()
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-    debugV("EffectManager::NextPalette");
-    g().CyclePalette(1);
-}
-
-void EffectManager::PreviousPalette()
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-    debugV("EffectManager::PreviousPalette");
-    g().CyclePalette(-1);
-}
-
-void EffectManager::LoadDefaultEffects()
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    _effectSetHashString = g_ptrEffectFactories->HashString();
-
-    for (const auto &numberedFactory : g_ptrEffectFactories->GetDefaultFactories())
-    {
-        auto pEffect = numberedFactory.CreateEffect();
-        if (pEffect)
-        {
-            // Effects in the default list are core effects. These can be disabled but not deleted.
-            pEffect->MarkAsCoreEffect();
-            _vEffects.push_back(pEffect);
-        }
-    }
-
-    SetInterval(DEFAULT_EFFECT_INTERVAL, true);
-
-    construct(true);
-}
-
-
 // Saves the current effect index to its own file. Note that this function only flags the writer
 //   for execution in the background.
 void EffectManager::SaveCurrentEffectIndex()
@@ -572,99 +221,6 @@ void EffectManager::SaveCurrentEffectIndex()
     if (g_ptrSystem->GetDeviceConfig().RememberCurrentEffect())
         // Default value for writer index is max value for size_t, so nothing will happen if writer has not yet been registered
         g_ptrSystem->GetJSONWriter().FlagWriter(l_CurrentEffectWriterIndex);
-}
-
-// Reads the current effect index from its own file. Returns true if the index was successfully read.
-bool EffectManager::ReadCurrentEffectIndex(size_t& index)
-{
-    File file = SPIFFS.open(CURRENT_EFFECT_CONFIG_FILE);
-    bool readIndex = false;
-
-    if (file)
-    {
-        if (file.size() > 0)
-        {
-            debugI("Attempting to read file %s", CURRENT_EFFECT_CONFIG_FILE);
-
-            auto valueString = file.readString();
-
-            if (!valueString.isEmpty())
-            {
-                index = strtoul(valueString.c_str(), nullptr, 10);
-                readIndex = true;
-            }
-        }
-
-        file.close();
-    }
-
-    return readIndex;
-}
-
-void EffectManager::LoadJSONEffects(const JsonArrayConst& effectsArray)
-{
-    std::set<int> loadedEffectNumbers;
-
-    // Create effects from JSON objects, using the respective factories in g_EffectFactories
-    auto& jsonFactories = g_ptrEffectFactories->GetJSONFactories();
-
-    for (auto effectObject : effectsArray)
-    {
-        int effectNumber = effectObject[PTY_EFFECTNR];
-        auto factoryEntry = jsonFactories.find(effectNumber);
-
-        if (factoryEntry == jsonFactories.end())
-            continue;
-
-        auto pEffect = factoryEntry->second(effectObject);
-        if (pEffect)
-        {
-            if (effectObject[PTY_COREEFFECT].as<int>())
-                pEffect->MarkAsCoreEffect();
-
-            _vEffects.push_back(pEffect);
-            loadedEffectNumbers.insert(effectNumber);
-        }
-    }
-}
-
-// Creates a copy of an existing effect in the list. Note that the effect is created but not yet added to the effect list;
-//   use the AppendEffect() function for that.
-std::shared_ptr<LEDStripEffect> EffectManager::CopyEffect(size_t index)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (index >= _vEffects.size())
-    {
-        debugW("Invalid index for CopyEffect");
-        return nullptr;
-    }
-
-    auto& sourceEffect = _vEffects[index];
-
-    const auto& jsonEffectFactories = g_ptrEffectFactories->GetJSONFactories();
-    auto factoryEntry = jsonEffectFactories.find(static_cast<int>(sourceEffect->effectId()));
-
-    if (factoryEntry == jsonEffectFactories.end())
-        return nullptr;
-
-    auto jsonDoc = CreateJsonDocument();
-    auto jsonObject = jsonDoc.to<JsonObject>();
-
-    if (!sourceEffect->SerializeToJSON(jsonObject))
-    {
-        debugE("Could not serialize effect %s to JSON", sourceEffect->FriendlyName().c_str());
-        return nullptr;
-    }
-
-    auto copiedEffect = factoryEntry->second(jsonDoc.as<JsonObjectConst>());
-
-    if (!copiedEffect)
-        return nullptr;
-
-    copiedEffect->SetEnabled(false);
-
-    return copiedEffect;
 }
 
 //
@@ -694,596 +250,6 @@ void RemoveEffectManagerConfig()
 
 void WriteCurrentEffectIndexFile()
 {
-    // Capture the current effect index without holding g_render_mutex to avoid nested lock order issues.
-    size_t currentIndex = g_ptrSystem->GetEffectManager().GetCurrentEffectIndex();
-
-    std::lock_guard renderPause(g_render_mutex);
-
-    #if USE_HUB75
-        HUB75GFX::WaitForMatrixSwap();
-    #endif
-
-    SPIFFS.remove(CURRENT_EFFECT_CONFIG_FILE);
-
-    File file = SPIFFS.open(CURRENT_EFFECT_CONFIG_FILE, FILE_WRITE);
-
-    if (!file)
-    {
-        debugE("Unable to open file %s for writing!", CURRENT_EFFECT_CONFIG_FILE);
-        return;
-    }
-
-    auto bytesWritten = file.print(currentIndex);
-
-    file.flush();
-    file.close();
-
-    debugI("Number of bytes written to file %s: %zu", CURRENT_EFFECT_CONFIG_FILE, (size_t)bytesWritten);
-
-    if (bytesWritten == 0)
-    {
-        debugE("Unable to write to file %s!", CURRENT_EFFECT_CONFIG_FILE);
-        SPIFFS.remove(CURRENT_EFFECT_CONFIG_FILE);
-    }
-}
-
-//
-// Other helper functions
-//
-
-#if ENABLE_AUDIO
-
-#include "effects/matrix/spectrumeffects.h"
-
-// GetSpectrumAnalyzer
-//
-// A little factory that makes colored spectrum analyzers
-
-std::shared_ptr<LEDStripEffect> GetSpectrumAnalyzer(CRGB color)
-{
-    CHSV hueColor = rgb2hsv_approximate(color);
-    CRGB color2 = CRGB(CHSV(hueColor.hue + 64, 255, 255));
-    auto object = make_shared_psram<SpectrumAnalyzerEffect>("Spectrum Clr", 24, CRGBPalette16(color, color2), true);
-    if (object->Init(g_ptrSystem->GetDevices()))
-        return object;
-    throw std::runtime_error("Could not initialize new spectrum analyzer, one color version!");
-}
-
-#endif
-
-#include "effects/strip/fireeffect.h"
-
-bool EffectManager::Init()
-{
-    for (const auto & _vEffect : _vEffects)
-    {
-        debugV("About to init effect %s", _vEffect->FriendlyName().c_str());
-        if (false == _vEffect->Init(_gfx))
-        {
-            debugW("Could not initialize effect: %s\n", _vEffect->FriendlyName().c_str());
-            return false;
-        }
-        debugV("Loaded Effect: %s", _vEffect->FriendlyName().c_str());
-    }
-    if (_vEffects.empty())
-        debugV("No local effects loaded");
-    else
-        debugV("First Effect: %s", GetCurrentEffectName().c_str());
-
-    if (g_ptrSystem->GetDeviceConfig().ApplyGlobalColors())
-        ApplyGlobalPaletteColors();
-
-    return true;
-}
-
-bool EffectManager::ReinitializeEffects()
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (!Init())
-        return false;
-
-    if (_tempEffect)
-    {
-        debugV("About to re-init temp effect %s", _tempEffect->FriendlyName().c_str());
-        if (!_tempEffect->Init(_gfx))
-        {
-            debugW("Could not re-initialize temporary effect: %s\n", _tempEffect->FriendlyName().c_str());
-            return false;
-        }
-    }
-
-    StartEffect();
-    return true;
-}
-
-bool EffectManager::ShowVU(bool bShow)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    auto& deviceConfig = g_ptrSystem->GetDeviceConfig();
-    bool bResult = deviceConfig.ShowVUMeter();
-    debugI("Setting ShowVU to %d\n", bShow);
-    deviceConfig.SetShowVUMeter(bShow);
-
-    // Erase any exising pixels since effects don't all clear each frame
-    if (!bShow)
-        _gfx[0]->setPixelsF(0, _gfx[0]->GetMatrixWidth(), CRGB::Black);
-
-    return bResult;
-}
-
-bool EffectManager::IsVUVisible() const
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-    return g_ptrSystem->GetDeviceConfig().ShowVUMeter() &&
-           (_tempEffect || !_vEffects.empty()) &&
-           GetCurrentEffect().CanDisplayVUMeter();
-}
-
-
-void EffectManager::ClearRemoteColor(bool retainRemoteEffect)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (!retainRemoteEffect)
-        _tempEffect = nullptr;
-
-    #if USE_HUB75
-        g().PausePalette(false);
-    #endif
-
-    g_ptrSystem->GetDeviceConfig().ClearApplyGlobalColors();
-}
-
-// ApplyGlobalColor
-//
-// When a global color is set via the remote, we create a fill effect and assign it as the "remote effect"
-// which takes drawing precedence
-
-void EffectManager::ApplyGlobalColor(CRGB color)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    debugI("Setting Global Color: %08lX\n", (unsigned long)(uint32_t)color);
-
-    auto& deviceConfig = g_ptrSystem->GetDeviceConfig();
-    deviceConfig.SetColorSettings(color, deviceConfig.GlobalColor());
-
-    ApplyGlobalPaletteColors();
-}
-
-void EffectManager::ApplyGlobalPaletteColors()
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    #if USE_HUB75
-        auto& pMatrix = g();
-        auto& deviceConfig = g_ptrSystem->GetDeviceConfig();
-        auto& globalColor = deviceConfig.GlobalColor();
-        auto& secondColor = deviceConfig.SecondColor();
-
-        // If the two colors are the same, we just shift the palette by 64 degrees to create a palette
-        // based from where those colors sit on the spectrum
-        if (secondColor == globalColor)
-        {
-            CHSV hsv = rgb2hsv_approximate(globalColor);
-            pMatrix.setPalette(CRGBPalette16(globalColor, CRGB(CHSV(hsv.hue + 64, 255, 255))));
-        }
-        else
-        {
-            // But if we have two different colors, we create a palette spread between them
-            pMatrix.setPalette(CRGBPalette16(secondColor, globalColor));
-        }
-
-        pMatrix.PausePalette(true);
-    #endif
-}
-
-void EffectManager::construct(bool clearTempEffect)
-{
-    _bPlayAll = false;
-
-    if (clearTempEffect && _tempEffect)
-        _clearTempEffectWhenExpired = true;
-
-    if (_vEffects.empty())
-    {
-        _iCurrentEffect = 0;
-        return;
-    }
-
-    if (_iCurrentEffect >= _vEffects.size())
-        _iCurrentEffect = _vEffects.size() - 1;
-
-    if (clearTempEffect && _tempEffect)
-    {
-        // This ensures that we start the correct effect after the temporary one.
-        //   The switching to the next effect is taken care of by NextEffect(), which starts with
-        //   increasing _iCurrentEffect. We therefore need to set it to the previous effect, to
-        //   make sure that the first effect after the temporary one is the one we want (either the
-        //   then current one when the chip was powered off, or the one at index 0).
-        if (_iCurrentEffect == 0)
-            _iCurrentEffect = EffectCount();
-
-        _iCurrentEffect--;
-    }
-}
-
-// DeserializeFromJSON
-//
-// This function deserializes LED strip effects from a provided JSON object.
-//
-// It first clears any existing effects and then attempts to populate the effects vector from
-// the provided JSON object, which should contain an array of effects configurations ("efs").
-//
-// For each effect in the JSON array, it attempts to create an effect from its JSON configuration.
-// If an effect is successfully created, it's added to the effects vector.
-//
-// If no effects are successfully loaded from JSON, it loads the default effects.
-//
-// If the JSON object includes an "eef" array, the function attempts to load each effect's enabled
-// state from it.
-// If the index exceeds the "eef" array's size, the effect is enabled by default.
-//
-// The function also sets the effect interval from the "ivl" field in the JSON object, defaulting
-// to a pre-defined value if the field isn't present.
-//
-// If the JSON object includes a "cei" field, the function sets the current effect index to this
-// value. If the value is greater than or equal to the number of effects, it defaults to the last
-// effect in the vector.
-//
-// Lastly, the function calls the construct() method, indicating successful deserialization.
-
-bool EffectManager::DeserializeFromJSON(const JsonObjectConst& jsonObject)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    ClearEffects();
-
-    // "efs" is the array of serialized effect objects
-    JsonArrayConst effectsArray = jsonObject["efs"].as<JsonArrayConst>();
-
-    // Check if the object actually contained an effect config array. If not, load the default effects.
-    if (effectsArray.isNull())
-    {
-        LoadDefaultEffects();
-        return true;
-    }
-
-    // Check if there's a persisted effect set version, and remember it if so
-    if (jsonObject[PTY_EFFECTSETVER].is<String>())
-        _effectSetHashString = jsonObject[PTY_EFFECTSETVER].as<String>();
-
-    LoadJSONEffects(effectsArray);
-
-    // If no effects were successfully loaded from JSON, it loads the default effects.
-    if (_vEffects.empty())
-        LoadDefaultEffects();
-
-    // "eef" was the array of effect enabled flags. They have now been integrated in the effects themselves;
-    //   this code is there to "migrate" users who already had a serialized effect config on their device
-    if (jsonObject["eef"].is<JsonArrayConst>())
-    {
-        // Try to load effect enabled state from JSON also, default to "enabled" otherwise
-        JsonArrayConst enabledArray = jsonObject["eef"].as<JsonArrayConst>();
-        size_t enabledSize = enabledArray.isNull() ? 0 : enabledArray.size();
-
-        for (int i = 0; i < _vEffects.size(); i++)
-        {
-            if (i >= enabledSize || enabledArray[i] == 1)
-                EnableEffect(i, true);
-            else
-                DisableEffect(i, true);
-        }
-    }
-
-    // "ivl" contains the effect interval in ms
-    SetInterval(jsonObject["ivl"].is<uint>() ? jsonObject["ivl"] : DEFAULT_EFFECT_INTERVAL, true);
-
-    // Try to read the effectindex from its own file. If that fails, "cei" may contain the current effect index instead
-    if (!ReadCurrentEffectIndex(_iCurrentEffect) && jsonObject["cei"].is<size_t>())
-        _iCurrentEffect = jsonObject["cei"];
-
-    // Make sure that if we read an index, it's sane
-    if (_vEffects.empty())
-        _iCurrentEffect = 0;
-    else if (_iCurrentEffect >= EffectCount())
-        _iCurrentEffect = EffectCount() - 1;
-
-    construct(true);
-
-    return true;
-}
-
-// SerializeToJSON - Serialize effects to a JSON object.
-//
-// This function serializes the current state of the LED strip effects into a JSON object.
-// It starts by setting the JSON format version ("PTY_VERSION") to a predefined value ("JSON_FORMAT_VERSION")
-// that helps in detecting and managing potential future incompatible structural updates.
-//
-// The function then sets the "ivl" and "cei" fields in the JSON object to the current effect interval
-// and the current effect index, respectively.
-//
-// Next, the function creates a nested array ("efs") in the JSON object to store the effects themselves.
-// It iterates through all effects, and for each effect, it creates a nested object in the effects array
-// and attempts to serialize the effect into this object. If serialization of any effect fails, the function
-// immediately returns false.
-//
-// If all effects are successfully serialized, the function returns true, indicating successful serialization.
-
-bool EffectManager::SerializeToJSON(JsonObject& jsonObject)
-{
-    std::lock_guard effectGuard(g_effect_manager_mutex);
-
-    // Set JSON format version to be able to detect and manage future incompatible structural updates
-    jsonObject[PTY_VERSION] = JSON_FORMAT_VERSION;
-    jsonObject["ivl"] = _effectInterval;
-    jsonObject[PTY_PROJECT] = PROJECT_NAME;
-    jsonObject[PTY_EFFECTSETVER] = _effectSetHashString;
-
-    // Next, the function creates a nested array ("efs") in the JSON object to store the effects themselves.
-    JsonArray effectsArray = jsonObject["efs"].to<JsonArray>();
-
-    for (auto & effect : _vEffects)
-    {
-        JsonObject effectObject = effectsArray.add<JsonObject>();
-        if (!(effect->SerializeToJSON(effectObject)))
-            return false;
-    }
-
-    return true;
-}
-
-void EffectManager::EnableEffect(size_t i, bool skipSave)
-{
-    // Web, remote, and render tasks all touch the effect list/current effect.
-    // Mutations take both locks so a UI request cannot reorder/delete/settings-
-    // mutate an effect while the draw loop is inside that effect's Draw().
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (i >= _vEffects.size())
-    {
-        debugW("Invalid index for EnableEffect");
-        return;
-    }
-
-    auto& effect = _vEffects[i];
-
-    if (!effect->IsEnabled())
-    {
-        if (!AreEffectsEnabled())
-            ClearRemoteColor(true);
-
-        effect->SetEnabled(true);
-
-        if (!skipSave)
-            SaveEffectManagerConfig();
-
-        {
-            std::lock_guard listenerGuard(_listenerMutex);
-            INFORM_EVENT_LISTENERS(_effectEventListeners, IEffectEventListener::OnEffectEnabledStateChanged, i, true);
-        }
-    }
-}
-
-void EffectManager::DisableEffect(size_t i, bool skipSave)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (i >= _vEffects.size())
-    {
-        debugW("Invalid index for DisableEffect");
-        return;
-    }
-
-    auto effect = _vEffects[i];
-
-    if (effect->IsEnabled())
-    {
-        effect->SetEnabled(false);
-
-        if (!AreEffectsEnabled())
-            ApplyGlobalColor(CRGB::Black);
-
-        if (!skipSave)
-            SaveEffectManagerConfig();
-
-        {
-            std::lock_guard listenerGuard(_listenerMutex);
-            INFORM_EVENT_LISTENERS(_effectEventListeners, IEffectEventListener::OnEffectEnabledStateChanged, i, false);
-        }
-    }
-}
-
-void EffectManager::MoveEffect(size_t from, size_t to)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (from >= _vEffects.size() || to >= _vEffects.size())
-    {
-        debugW("Invalid index for MoveEffect");
-        return;
-    }
-
-    if (from == to)
-        return;
-    else if (from < to)
-        std::rotate(_vEffects.begin() + from, _vEffects.begin() + from + 1, _vEffects.begin() + to + 1);
-    else // from > to
-        std::rotate(_vEffects.rend() - from - 1, _vEffects.rend() - from, _vEffects.rend() - to);
-
-    if (from == _iCurrentEffect)
-    {
-        _iCurrentEffect = to;
-        SaveCurrentEffectIndex();
-    }
-    else if (from < _iCurrentEffect && to >= _iCurrentEffect)
-    {
-        _iCurrentEffect--;
-        SaveCurrentEffectIndex();
-    }
-    else if (from > _iCurrentEffect && to <= _iCurrentEffect)
-    {
-        _iCurrentEffect++;
-        SaveCurrentEffectIndex();
-    }
-
-    SaveEffectManagerConfig();
-
-    {
-        std::lock_guard listenerGuard(_listenerMutex);
-        INFORM_EVENT_LISTENERS(_effectEventListeners, IEffectEventListener::OnEffectListDirty);
-    }
-}
-
-// Adds an effect to the effect list and enables it. If an effect is added that is already in the effect list then the result
-//   is undefined but potentially messy.
-bool EffectManager::AppendEffect(std::shared_ptr<LEDStripEffect>& effect)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (!effect->Init(_gfx))
-        return false;
-
-    _vEffects.push_back(effect);
-    EnableEffect(_vEffects.size() - 1, true);
-
-    SaveEffectManagerConfig();
-
-    {
-        std::lock_guard listenerGuard(_listenerMutex);
-        INFORM_EVENT_LISTENERS(_effectEventListeners, IEffectEventListener::OnEffectListDirty);
-    }
-
-    return true;
-}
-
-// Deletes an effect from the effect list. Note that core effects cannot be deleted.
-bool EffectManager::DeleteEffect(size_t index)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (index >= _vEffects.size())
-    {
-        debugW("Invalid index for DeleteEffect");
-        return false;
-    }
-
-    if (_vEffects[index]->IsCoreEffect())
-        return false;
-
-    DisableEffect(index, true);
-
-    if (index == _iCurrentEffect && _vEffects.size() > 1)
-        NextEffect(true);
-
-    _vEffects.erase(_vEffects.begin() + index);
-
-    if (_vEffects.empty())
-    {
-        _iCurrentEffect = 0;
-    }
-    else if (index <= _iCurrentEffect)
-    {
-        if (_iCurrentEffect > 0)
-            _iCurrentEffect--;
-        else if (_iCurrentEffect >= _vEffects.size())
-            _iCurrentEffect = _vEffects.size() - 1;
-    }
-
-    SaveCurrentEffectIndex();
-    SaveEffectManagerConfig();
-
-    {
-        std::lock_guard listenerGuard(_listenerMutex);
-        INFORM_EVENT_LISTENERS(_effectEventListeners, IEffectEventListener::OnEffectListDirty);
-    }
-
-    return true;
-}
-
-void EffectManager::CheckEffectTimerExpired()
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (!_tempEffect && _vEffects.empty())
-        return;
-
-    if (IsIntervalEternal() && !GetCurrentEffect().HasMaximumEffectTime())
-        return;
-
-    if (GetTimeUsedByCurrentEffect() >= GetEffectiveInterval())
-    {
-        if (_clearTempEffectWhenExpired)
-        {
-            _tempEffect.reset();
-            _clearTempEffectWhenExpired = false;
-        }
-
-        debugV("%ldms elapsed: Next Effect", millis() - _effectStartTime);
-        NextEffect();
-        debugV("Current Effect: %s", GetCurrentEffectName().c_str());
-    }
-}
-
-// Update to the next effect and abort the current effect.
-
-void EffectManager::NextEffect(bool skipSave)
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (_vEffects.empty())
-    {
-        _iCurrentEffect = 0;
-        _effectStartTime = millis();
-        return;
-    }
-
-    auto enabled = AreEffectsEnabled();
-
-    do
-    {
-        _iCurrentEffect++;
-        _iCurrentEffect %= EffectCount();
-        _effectStartTime = millis();
-    } while (enabled && false == _bPlayAll && false == IsEffectEnabled(_iCurrentEffect));
-
-    StartEffect();
-    if (!skipSave)
-        SaveCurrentEffectIndex();
-
-    {
-        std::lock_guard listenerGuard(_listenerMutex);
-        INFORM_EVENT_LISTENERS(_effectEventListeners, IEffectEventListener::OnCurrentEffectChanged, _iCurrentEffect);
-    }
-}
-
-// Go back to the previous effect and abort the current one.
-
-void EffectManager::PreviousEffect()
-{
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-    if (_vEffects.empty())
-    {
-        _iCurrentEffect = 0;
-        _effectStartTime = millis();
-        return;
-    }
-
-    auto enabled = AreEffectsEnabled();
-
-    do
-    {
-        if (_iCurrentEffect == 0)
-            _iCurrentEffect = EffectCount();
-
-        _iCurrentEffect--;
-        _effectStartTime = millis();
-    } while (enabled && false == _bPlayAll && false == IsEffectEnabled(_iCurrentEffect));
-
-    StartEffect();
     SaveCurrentEffectIndex();
 
     {
