+++ src/gfxbase.cpp
@@ -537,962 +537,6 @@ void GFXBase::BlurFrame(int amount)
     blur2d(leds, _width, 0, _height, 1, amount);
 }
 
-void GFXBase::CyclePalette(int offset)
-{
-    loadPalette(_paletteIndex + offset);
-}
-
-void GFXBase::ChangePalettePeriodically()
-{
-    if (_palettePaused)
-        return;
-
-    const int minutesPerPaletteCycle = 2;
-    uint8_t secondHand = ((millis() / minutesPerPaletteCycle) / 1000) % 60;
-
-    if (_lastSecond != secondHand)
-    {
-        _lastSecond = secondHand;
-        if (secondHand == 0)
-        {
-            _targetPalette = RainbowColors_p;
-        }
-        if (secondHand == 10)
-        {
-            _targetPalette = HeatColors_p;
-        } // CRGBPalette16( g,g,b,b, p,p,b,b, g,g,b,b, p,p,b,b); }
-        if (secondHand == 20)
-        {
-            _targetPalette = ForestColors_p;
-        } // CRGBPalette16( b,b,b,w, b,b,b,w, b,b,b,w, b,b,b,w); }
-        if (secondHand == 30)
-        {
-            _targetPalette = LavaColors_p;
-        } // Black gaps
-        if (secondHand == 40)
-        {
-            _targetPalette = CloudColors_p;
-        }
-        if (secondHand == 50)
-        {
-            _targetPalette = PartyColors_p;
-        }
-    }
-}
-
-// Cross-fade current palette slowly toward the target palette
-//
-// Each time that nblendPaletteTowardPalette is called, small changes
-// are made to currentPalette to bring it closer to matching targetPalette.
-// You can control how many changes are made in each call:
-//   - the default of 24 is a good balance
-//   - meaningful values are 1-48.  1=very very slow, 48=quickest
-//   - "0" means do not change the currentPalette at all; freeze
-
-void GFXBase::PausePalette(bool bPaused)
-{
-    _palettePaused = bPaused;
-}
-
-void GFXBase::UpdatePaletteCycle()
-{
-    ChangePalettePeriodically();
-    uint8_t maxChanges = 24;
-    nblendPaletteTowardPalette(_currentPalette, _targetPalette, maxChanges);
-}
-
-void GFXBase::RandomPalette()
-{
-    loadPalette(_randomPaletteIndex);
-}
-
-void GFXBase::fillRectangle(int x0, int y0, int x1, int y1, CRGB color)
-{
-    for (int x = x0; x < x1; x++)
-        for (int y = y0; y < y1; y++)
-            drawPixel(x, y, color);
-}
-
-void GFXBase::setPalette(const CRGBPalette16& palette)
-{
-    _currentPalette = palette;
-    _targetPalette = palette;
-    _currentPaletteName = "Custom";
-}
-
-// loadPalette
-//
-// Note that this function may recurse without
-// bound if your random() is very very dumb.
-
-void GFXBase::loadPalette(int index)
-{
-    _paletteIndex = index;
-
-    if (_paletteIndex >= _paletteCount)
-        _paletteIndex = 0;
-    else if (_paletteIndex < 0)
-        _paletteIndex = _paletteCount - 1;
-
-    switch (_paletteIndex)
-    {
-    case 0:
-        _targetPalette = RainbowColors_p;
-        _currentPaletteName = "Rainbow";
-        break;
-    case 1:
-        _targetPalette = OceanColors_p;
-        _currentPaletteName = "Ocean";
-        break;
-    case 2:
-        _targetPalette = CloudColors_p;
-        _currentPaletteName = "Cloud";
-        break;
-    case 3:
-        _targetPalette = ForestColors_p;
-        _currentPaletteName = "Forest";
-        break;
-    case 4:
-        _targetPalette = PartyColors_p;
-        _currentPaletteName = "Party";
-        break;
-    case 5:
-        setupGrayscalePalette();
-        _currentPaletteName = "Grey";
-        break;
-    case _heatColorsPaletteIndex:
-        _targetPalette = HeatColors_p;
-        _currentPaletteName = "Heat";
-        break;
-    case 7:
-        _targetPalette = LavaColors_p;
-        _currentPaletteName = "Lava";
-        break;
-    case 8:
-        setupIcePalette();
-        _currentPaletteName = "Ice";
-        break;
-    case _randomPaletteIndex:
-        loadPalette(random(0, _paletteCount - 1));
-        _paletteIndex = _randomPaletteIndex;
-        _currentPaletteName = "Random";
-        break;
-    }
-    _currentPalette = _targetPalette;
-}
-
-void GFXBase::setPalette(const String& paletteName)
-{
-    static const std::unordered_map<const char*, int> paletteMap = {
-        {"Rainbow", 0},
-        {"Ocean", 1},
-        {"Cloud", 2},
-        {"Forest", 3},
-        {"Party", 4},
-        {"Grayscale", 5},
-        {"Heat", 6},
-        {"Lava", 7},
-        {"Ice", 8}
-    };
-
-    auto it = paletteMap.find(paletteName.c_str());
-    if (it != paletteMap.end()) {
-        loadPalette(it->second);
-    } else if (paletteName == "Random") {
-        RandomPalette();
-    }
-}
-
-void GFXBase::listPalettes()
-{
-    Serial.println("{");
-    Serial.print("  \"count\": ");
-    Serial.print(_paletteCount);
-    Serial.println(",");
-    Serial.println("  \"results\": [");
-
-    static constexpr const char* paletteNames[] =
-    {
-        "Rainbow", "Ocean", "Cloud", "Forest", "Party",
-        "Grayscale", "Heat", "Lava", "Ice", "Random"
-    };
-
-    for (int i = 0; i < _paletteCount; i++)
-    {
-        Serial.print("    \"");
-        Serial.print(paletteNames[i]);
-        if (i == _paletteCount - 1)
-            Serial.println("\"");
-        else
-            Serial.println("\",");
-    }
-
-    Serial.println("  ]");
-    Serial.println("}");
-}
-
-void GFXBase::setupGrayscalePalette()
-{
-    _targetPalette = CRGBPalette16(CRGB::Black, CRGB::White);
-}
-
-void GFXBase::setupIcePalette()
-{
-    _targetPalette = CRGBPalette16(CRGB::Black, CRGB::Blue, CRGB::Aqua, CRGB::White);
-}
-
-// Oscillators and Emitters
-
-// set the speeds (and by that ratios) of the oscillators here
-
-void GFXBase::MoveOscillators()
-{
-    osci[0] = osci[0] + 5;
-    osci[1] = osci[1] + 2;
-    osci[2] = osci[2] + 3;
-    osci[3] = osci[3] + 4;
-    osci[4] = osci[4] + 1;
-    if (osci[4] % 2 == 0)
-        osci[5] = osci[5] + 1; // .5
-    for (int i = 0; i < 4; i++)
-    {
-        p[i] = map8(sin8(osci[i]), 0, std::min(255U, (unsigned int)_width - 1)); // why? to keep the result in the range of 0-_width (matrix size)
-    }
-}
-
-void GFXBase::ResetOscillators()
-{
-    std::fill_n(osci, 6, 0);
-    std::fill_n(p, 6, 0);
-}
-
-// All the Caleidoscope functions work directly within the screenbuffer (leds array).
-// Draw whatever you like in the area x(0-15) and y (0-15) and then copy it around.
-
-// rotates the first 16x16 quadrant 3 times onto a 32x32 (+90 degrees rotation for each one)
-
-void GFXBase::Caleidoscope1() const
-{
-    for (int x = 0; x < ((_width + 1) / 2); x++)
-    {
-        for (int y = 0; y < ((_height + 1) / 2); y++)
-        {
-            leds[XY(_width - 1 - x, y)] = leds[XY(x, y)];
-            leds[XY(_width - 1 - x, _height - 1 - y)] = leds[XY(x, y)];
-            leds[XY(x, _height - 1 - y)] = leds[XY(x, y)];
-        }
-    }
-}
-
-// mirror the first 16x16 quadrant 3 times onto a 32x32
-
-void GFXBase::Caleidoscope2() const
-{
-    for (int x = 0; x < ((_width + 1) / 2); x++)
-    {
-        for (int y = 0; y < ((_height + 1) / 2); y++)
-        {
-            leds[XY(_width - 1 - x, y)] = leds[XY(y, x)];
-            leds[XY(x, _height - 1 - y)] = leds[XY(y, x)];
-            leds[XY(_width - 1 - x, _height - 1 - y)] = leds[XY(x, y)];
-        }
-    }
-}
-
-// copy one diagonal triangle into the other one within a 16x16
-
-void GFXBase::Caleidoscope3() const
-{
-    for (int x = 0; x < ((_width + 1) / 2); x++)
-    {
-        for (int y = 0; y <= x; y++)
-        {
-            leds[XY(x, y)] = leds[XY(y, x)];
-        }
-    }
-}
-
-// copy one diagonal triangle into the other one within a 16x16 (90 degrees rotated compared to Caleidoscope3)
-
-void GFXBase::Caleidoscope4() const
-{
-    for (int x = 0; x < ((_width + 1) / 2); x++)
-    {
-        for (int y = 0; y <= ((_height + 1) / 2) - x; y++)
-        {
-            leds[XY(((_height + 1) / 2) - y, ((_width + 1) / 2) - x)] = leds[XY(x, y)];
-        }
-    }
-}
-
-// copy one diagonal triangle into the other one within a 8x8
-
-void GFXBase::Caleidoscope5() const
-{
-    for (int x = 0; x < _width / 4; x++)
-    {
-        for (int y = 0; y <= x; y++)
-        {
-            leds[XY(x, y)] = leds[XY(y, x)];
-        }
-    }
-
-    for (int x = _width / 4; x < _width / 2; x++)
-    {
-        for (int y = _height / 4; y >= 0; y--)
-        {
-            leds[XY(x, y)] = leds[XY(y, x)];
-        }
-    }
-}
-
-// rotates the first 8x8 quadrant 3 times onto a 16x16
-
-void GFXBase::Caleidoscope6() const
-{
-    for (int x = 1; x < ((_width + 1) / 2); x++)
-    {
-        leds[XY(7 - x, 7)] = leds[XY(x, 0)];
-    } // a
-    for (int x = 2; x < ((_width + 1) / 2); x++)
-    {
-        leds[XY(7 - x, 6)] = leds[XY(x, 1)];
-    } // b
-    for (int x = 3; x < ((_width + 1) / 2); x++)
-    {
-        leds[XY(7 - x, 5)] = leds[XY(x, 2)];
-    } // c
-    for (int x = 4; x < ((_width + 1) / 2); x++)
-    {
-        leds[XY(7 - x, 4)] = leds[XY(x, 3)];
-    } // d
-    for (int x = 5; x < ((_width + 1) / 2); x++)
-    {
-        leds[XY(7 - x, 3)] = leds[XY(x, 4)];
-    } // e
-    for (int x = 6; x < ((_width + 1) / 2); x++)
-    {
-        leds[XY(7 - x, 2)] = leds[XY(x, 5)];
-    } // f
-    for (int x = 7; x < ((_width + 1) / 2); x++)
-    {
-        leds[XY(7 - x, 1)] = leds[XY(x, 6)];
-    } // g
-}
-
-// SpiralStream
-//
-// create a square twister to the left or counter-clockwise
-// x and y for center, r for radius
-
-void GFXBase::SpiralStream(int x, int y, int r, uint8_t dimm) const
-{
-    for (int d = r; d >= 0; d--)
-    { // from the outside to the inside
-        for (int i = x - d; i <= x + d; i++)
-        {
-            leds[XY(i, y - d)] += leds[XY(i + 1, y - d)]; // lowest row to the right
-            leds[XY(i, y - d)].nscale8(dimm);
-        }
-        for (int i = y - d; i <= y + d; i++)
-        {
-            leds[XY(x + d, i)] += leds[XY(x + d, i + 1)]; // right colum up
-            leds[XY(x + d, i)].nscale8(dimm);
-        }
-        for (int i = x + d; i >= x - d; i--)
-        {
-            leds[XY(i, y + d)] += leds[XY(i - 1, y + d)]; // upper row to the left
-            leds[XY(i, y + d)].nscale8(dimm);
-        }
-        for (int i = y + d; i >= y - d; i--)
-        {
-            leds[XY(x - d, i)] += leds[XY(x - d, i - 1)]; // left colum down
-            leds[XY(x - d, i)].nscale8(dimm);
-        }
-    }
-}
-
-// expand everything within a circle
-
-void GFXBase::Expand(int centerX, int centerY, int radius, uint8_t dimm)
-{
-    if (radius == 0)
-        return;
-
-    int currentRadius = radius;
-
-    while (currentRadius > 0)
-    {
-        int a = radius, b = 0;
-        int radiusError = 1 - a;
-
-        int nextRadius = currentRadius - 1;
-        int nextA = nextRadius - 1, nextB = 0;
-        int nextRadiusError = 1 - nextA;
-
-        while (a >= b)
-        {
-            // move them out one pixel on the radius
-            leds[XY(a + centerX, b + centerY)]   = leds[XY(nextA + centerX, nextB + centerY)];
-            leds[XY(b + centerX, a + centerY)]   = leds[XY(nextB + centerX, nextA + centerY)];
-            leds[XY(-a + centerX, b + centerY)]  = leds[XY(-nextA + centerX, nextB + centerY)];
-            leds[XY(-b + centerX, a + centerY)]  = leds[XY(-nextB + centerX, nextA + centerY)];
-            leds[XY(-a + centerX, -b + centerY)] = leds[XY(-nextA + centerX, -nextB + centerY)];
-            leds[XY(-b + centerX, -a + centerY)] = leds[XY(-nextB + centerX, -nextA + centerY)];
-            leds[XY(a + centerX, -b + centerY)]  = leds[XY(nextA + centerX, -nextB + centerY)];
-            leds[XY(b + centerX, -a + centerY)]  = leds[XY(nextB + centerX, -nextA + centerY)];
-
-            // dim them
-            leds[XY(a + centerX, b + centerY)].nscale8(dimm);
-            leds[XY(b + centerX, a + centerY)].nscale8(dimm);
-            leds[XY(-a + centerX, b + centerY)].nscale8(dimm);
-            leds[XY(-b + centerX, a + centerY)].nscale8(dimm);
-            leds[XY(-a + centerX, -b + centerY)].nscale8(dimm);
-            leds[XY(-b + centerX, -a + centerY)].nscale8(dimm);
-            leds[XY(a + centerX, -b + centerY)].nscale8(dimm);
-            leds[XY(b + centerX, -a + centerY)].nscale8(dimm);
-
-            b++;
-            if (radiusError < 0)
-                radiusError += 2 * b + 1;
-            else
-            {
-                a--;
-                radiusError += 2 * (b - a + 1);
-            }
-
-            nextB++;
-            if (nextRadiusError < 0)
-                nextRadiusError += 2 * nextB + 1;
-            else
-            {
-                nextA--;
-                nextRadiusError += 2 * (nextB - nextA + 1);
-            }
-        }
-
-        currentRadius--;
-    }
-}
-
-// give it a linear tail to the right
-
-void GFXBase::StreamRight(uint8_t scale, int fromX, int toX, int fromY, int toY)
-{
-    for (int x = fromX + 1; x < toX; x++)
-    {
-        for (int y = fromY; y < toY; y++)
-        {
-            leds[XY(x, y)] += leds[XY(x - 1, y)];
-            leds[XY(x, y)].nscale8(scale);
-        }
-    }
-    for (int y = fromY; y < toY; y++)
-        leds[XY(0, y)].nscale8(scale);
-}
-
-// give it a linear tail to the left
-
-void GFXBase::StreamLeft(uint8_t scale, int fromX, int toX, int fromY, int toY)
-{
-    for (int x = toX; x < fromX; x++)
-    {
-        for (int y = fromY; y < toY; y++)
-        {
-            leds[XY(x, y)] += leds[XY(x + 1, y)];
-            leds[XY(x, y)].nscale8(scale);
-        }
-    }
-    for (int y = fromY; y < toY; y++)
-        leds[XY(0, y)].nscale8(scale);
-}
-
-// give it a linear tail downwards
-
-void GFXBase::StreamDown(uint8_t scale)
-{
-    for (int x = 0; x < _width; x++)
-    {
-        for (int y = 1; y < _height; y++)
-        {
-            leds[XY(x, y)] += leds[XY(x, y - 1)];
-            leds[XY(x, y)].nscale8(scale);
-        }
-    }
-    for (int x = 0; x < _width; x++)
-        leds[XY(x, 0)].nscale8(scale);
-}
-
-// give it a linear tail upwards
-
-void GFXBase::StreamUp(uint8_t scale)
-{
-    for (int x = 0; x < _width; x++)
-    {
-        for (int y = (int)_height - 2; y >= 0; y--)
-        {
-            leds[XY(x, y)] += leds[XY(x, y + 1)];
-            leds[XY(x, y)].nscale8(scale);
-        }
-    }
-    for (int x = 0; x < _width; x++)
-        leds[XY(x, (int)_height - 1)].nscale8(scale);
-}
-
-// give it a linear tail up and to the left
-
-void GFXBase::StreamUpAndLeft(uint8_t scale)
-{
-    for (int x = 0; x < (int)_width - 1; x++)
-    {
-        for (int y = (int)_height - 2; y >= 0; y--)
-        {
-            leds[XY(x, y)] += leds[XY(x + 1, y + 1)];
-            leds[XY(x, y)].nscale8(scale);
-        }
-    }
-    for (int x = 0; x < (int)_width; x++)
-        leds[XY(x, (int)_height - 1)].nscale8(scale);
-    for (int y = 0; y < (int)_height; y++)
-        leds[XY((int)_width - 1, y)].nscale8(scale);
-}
-
-// give it a linear tail up and to the right
-
-void GFXBase::StreamUpAndRight(uint8_t scale)
-{
-    for (int x = 0; x < (int)_width - 1; x++)
-    {
-        for (int y = (int)_height - 2; y >= 0; y--)
-        {
-            leds[XY(x + 1, y)] += leds[XY(x, y + 1)];
-            leds[XY(x, y)].nscale8(scale);
-        }
-    }
-    for (int x = 0; x < (int)_width; x++)
-        leds[XY(x, (int)_height - 1)].nscale8(scale);
-
-    for (int y = 0; y < (int)_height; y++)
-        leds[XY((int)_width - 1, y)].nscale8(scale);
-}
-
-// just move everything one line down
-
-void GFXBase::MoveDown()
-{
-    for (int y = (int)_height - 1; y > 0; y--)
-    {
-        for (int x = 0; x < (int)_width; x++)
-        {
-            leds[XY(x, y)] = leds[XY(x, y - 1)];
-        }
-    }
-}
-
-// just move everything one line down
-
-void GFXBase::VerticalMoveFrom(int start, int end)
-{
-    for (int y = end; y > start; y--)
-    {
-        for (int x = 0; x < (int)_width; x++)
-        {
-            leds[XY(x, y)] = leds[XY(x, y - 1)];
-        }
-    }
-}
-
-// copy the rectangle defined with 2 points x0, y0, x1, y1
-// to the rectangle beginning at x2, x3
-
-void GFXBase::Copy(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2)
-{
-    for (int y = y0; y < y1 + 1; y++)
-    {
-        for (int x = x0; x < x1 + 1; x++)
-        {
-            leds[XY(x + x2 - x0, y + y2 - y0)] = leds[XY(x, y)];
-        }
-    }
-}
-
-void GFXBase::BresenhamLine(int x0, int y0, int x1, int y1, CRGB color, bool bMerge)
-{
-    int dx = abs(x1 - x0); // Delta in x direction
-    int dy = abs(y1 - y0); // Delta in y direction
-    int sx = (x0 < x1) ? 1 : -1; // Step in x direction
-    int sy = (y0 < y1) ? 1 : -1; // Step in y direction
-
-    int err = dx - dy; // Initial error term
-
-    while (true)
-    {
-        int index = XY(x0, y0);
-        if (isValidPixel(index))
-        {
-            // Optimization opportunity: unswtitch bMerge into another function
-            leds[index] = bMerge ? leds[index] + color : color;
-        }
-
-        if (x0 == x1 && y0 == y1)
-            break; // Exit the loop once we've reached the destination
-
-        int e2 = 2 * err; // Error term multiplied by 2 for efficiency. Saves second test for Y.
-        if (e2 > -dy) // Move in the x direction if needed
-        {
-            err -= dy;
-            x0 += sx;
-        }
-
-        if (e2 < dx) // Move in the y direction if needed
-        {
-            err += dx;
-            y0 += sy;
-        }
-    }
-}
-
-void GFXBase::BresenhamLine(int x0, int y0, int x1, int y1, uint8_t colorIndex, bool bMerge)
-{
-    BresenhamLine(x0, y0, x1, y1, ColorFromCurrentPalette(colorIndex), bMerge);
-}
-
-void GFXBase::drawLine(int x0, int y0, int x1, int y1, CRGB color)
-{
-    BresenhamLine(x0, y0, x1, y1, color);
-}
-
-void GFXBase::DimAll(uint8_t value)
-{
-    for (int i = 0; i < (int)_ledcount; i++)
-        fadePixelToBlackBy(i, 255 - value);
-}
-
-CRGB GFXBase::ColorFromCurrentPalette(uint8_t index, uint8_t brightness, TBlendType blendType) const
-{
-    return ColorFromPalette(_currentPalette, index, brightness, _currentBlendType);
-}
-
-CRGB GFXBase::HsvToRgb(uint8_t h, uint8_t s, uint8_t v)
-{
-    CHSV hsv = CHSV(h, s, v);
-    CRGB rgb;
-    hsv2rgb_spectrum(hsv, rgb);
-    return rgb;
-}
-
-#if USE_NOISE
-    // The following functions are specializations of noise-related member function
-    // templates declared in gfxbase.h.
-
-    void GFXBase::SetNoise(uint32_t nx, uint32_t ny, uint32_t nz, uint32_t sx, uint32_t sy)
-    {
-        EnsureNoise();
-        _ptrNoise->noise_x += nx;
-        _ptrNoise->noise_y += ny;
-        _ptrNoise->noise_z += nz;
-        _ptrNoise->noise_scale_x = sx;
-        _ptrNoise->noise_scale_y = sy;
-    }
-
-    // Internal implementation: assumes _ptrNoise is already initialized.
-    // Must NOT call EnsureNoise() — this is invoked from within EnsureNoise()'s
-    // call_once lambda, and a recursive call_once on the same flag is UB.
-    void GFXBase::FillGetNoiseImpl() const
-    {
-        // Subtracting the center offset before scaling ensures the noise pattern radiates
-        // outwards from the center of the display (exactly as #803 intended).
-        //
-        // We use uint32_t for the indices as it's the native register size for the ESP32,
-        // avoiding the unnecessary overhead of masking/extending smaller types.
-        for (uint32_t i = 0; i < _width; i++)
-        {
-            int32_t ioffset = _ptrNoise->noise_scale_x * (int32_t)(i - (_width / 2));
-
-            for (uint32_t j = 0; j < _height; j++)
-            {
-                int32_t joffset = _ptrNoise->noise_scale_y * (int32_t)(j - (_height / 2));
-                uint8_t data    = inoise16(_ptrNoise->noise_x + ioffset, _ptrNoise->noise_y + joffset, _ptrNoise->noise_z) >> 8;
-                uint8_t olddata = _ptrNoise->noise[i][j];
-                uint8_t newdata = scale8(olddata, _ptrNoise->noisesmoothing) + scale8(data, 256 - _ptrNoise->noisesmoothing);
-
-                _ptrNoise->noise[i][j] = newdata;
-            }
-        }
-    }
-
-    void GFXBase::FillGetNoise() const
-    {
-        if (!EnsureNoise())
-            FillGetNoiseImpl();
-    }
-
-    template<>
-    void GFXBase::MoveFractionalNoiseX<NoiseApproach::MRI>(uint8_t amt, uint8_t shift)
-    {
-        EnsureNoise();
-        std::unique_ptr<CRGB[]> ledsTemp = std::make_unique<CRGB[]>(_ledcount);
-
-        // move delta pixelwise
-        for (uint32_t y = 0; y < _height; y++)
-        {
-            uint16_t amount = _ptrNoise->noise[0][y] * amt;
-            uint8_t delta = _width - 1 - (amount / 256);
-
-            // Process up to the end less the delta
-            for (uint32_t x = 0; x < _width - delta; x++)
-                ledsTemp[XY(x, y)] = leds[XY(x + delta, y)];
-
-            // Do the tail portion while wrapping around
-            for (uint32_t x = _width - delta; x < _width; x++)
-                ledsTemp[XY(x, y)] = leds[XY(x + delta - _width, y)];
-        }
-
-        // move fractions
-        CRGB PixelA;
-        CRGB PixelB;
-
-        for (uint32_t y = 0; y < _height; y++)
-        {
-            uint16_t amount = _ptrNoise->noise[0][y] * amt;
-            uint8_t delta = _width - 1 - (amount / 256);
-            uint8_t fractions = amount - (delta * 256);
-
-            for (uint32_t x = 1; x < _width; x++)
-            {
-                PixelA = ledsTemp[XY(x, y)];
-                PixelB = ledsTemp[XY(x - 1, y)];
-
-                PixelA %= 255 - fractions;
-                PixelB %= fractions;
-
-                leds[XY(x, y)] = PixelA + PixelB;
-            }
-
-            PixelA = ledsTemp[XY(0, y)];
-            PixelB = ledsTemp[XY(_width - 1, y)];
-
-            PixelA %= 255 - fractions;
-            PixelB %= fractions;
-
-            leds[XY(0, y)] = PixelA + PixelB;
-        }
-    }
-
-    template<>
-    void GFXBase::MoveFractionalNoiseX<NoiseApproach::General>(uint8_t amt, uint8_t shift)
-    {
-        EnsureNoise();
-        // Aligning with Approach::One while keeping the "Approach::Two" optimized behavior.
-        // We use int32_t for the 'amount' and 'delta' as they can be large or negative.
-        for (uint32_t y = 0; y < _height; y++)
-        {
-            int32_t amount = ((int32_t)_ptrNoise->noise[0][y] - 128) * 2 * amt + shift * 256;
-            int32_t delta = abs(amount) >> 8;
-            int32_t fraction = abs(amount) & 255;
-
-            for (uint32_t x = 0; x < _width; x++)
-            {
-                int32_t zD;
-                int32_t zF;
-
-                if (amount < 0)
-                {
-                    zD = (int32_t)x - delta;
-                    zF = zD - 1;
-                }
-                else
-                {
-                    zD = (int32_t)x + delta;
-                    zF = zD + 1;
-                }
-
-                CRGB PixelA = CRGB::Black;
-                if ((zD >= 0) && (zD < (int32_t)_width))
-                    PixelA = leds[XY(zD, y)];
-
-                CRGB PixelB = CRGB::Black;
-                if ((zF >= 0) && (zF < (int32_t)_width))
-                    PixelB = leds[XY(zF, y)];
-
-                leds[XY(x, y)] = (PixelA.nscale8(ease8InOutApprox(255 - fraction))) + (PixelB.nscale8(ease8InOutApprox(fraction)));
-            }
-        }
-    }
-
-    template<>
-    void GFXBase::MoveFractionalNoiseY<NoiseApproach::MRI>(uint8_t amt, uint8_t shift)
-    {
-        EnsureNoise();
-        std::unique_ptr<CRGB[]> ledsTemp = std::make_unique<CRGB[]>(_ledcount);
-
-        // move delta pixelwise
-        for (uint32_t x = 0; x < _width; x++)
-        {
-            uint16_t amount = _ptrNoise->noise[x][0] * amt;
-            uint8_t delta = _height - 1 - (amount / 256);
-
-            for (uint32_t y = 0; y < _height - delta; y++)
-            {
-                ledsTemp[XY(x, y)] = leds[XY(x, y + delta)];
-            }
-            for (uint32_t y = _height - delta; y < _height; y++)
-            {
-                ledsTemp[XY(x, y)] = leds[XY(x, y + delta - _height)];
-            }
-        }
-
-        // move fractions
-        CRGB PixelA;
-        CRGB PixelB;
-
-        for (uint32_t x = 0; x < _width; x++)
-        {
-            uint16_t amount = _ptrNoise->noise[x][0] * amt;
-            uint8_t delta = _height - 1 - (amount / 256);
-            uint8_t fractions = amount - (delta * 256);
-
-            for (uint32_t y = 1; y < _height; y++)
-            {
-                PixelA = ledsTemp[XY(x, y)];
-                PixelB = ledsTemp[XY(x, y - 1)];
-
-                PixelA %= 255 - fractions;
-                PixelB %= fractions;
-
-                leds[XY(x, y)] = PixelA + PixelB;
-            }
-
-            PixelA = ledsTemp[XY(x, 0)];
-            PixelB = ledsTemp[XY(x, _height - 1)];
-
-            PixelA %= 255 - fractions;
-            PixelB %= fractions;
-
-            leds[XY(x, 0)] = PixelA + PixelB;
-        }
-    }
-
-    template<>
-    void GFXBase::MoveFractionalNoiseY<NoiseApproach::General>(uint8_t amt, uint8_t shift)
-    {
-        EnsureNoise();
-        for (uint32_t x = 0; x < _width; x++)
-        {
-            int32_t amount = ((int32_t)_ptrNoise->noise[x][0] - 128) * 2 * amt + shift * 256;
-            int32_t delta = abs(amount) >> 8;
-            int32_t fraction = abs(amount) & 255;
-
-            for (uint32_t y = 0; y < _height; y++)
-            {
-                int32_t zD;
-                int32_t zF;
-
-                if (amount < 0)
-                {
-                    zD = (int32_t)y - delta;
-                    zF = zD - 1;
-                }
-                else
-                {
-                    zD = (int32_t)y + delta;
-                    zF = zD + 1;
-                }
-
-                CRGB PixelA = CRGB::Black;
-                if ((zD >= 0) && (zD < (int32_t)_height))
-                    PixelA = leds[XY(x, zD)];
-
-                CRGB PixelB = CRGB::Black;
-                if ((zF >= 0) && (zF < (int32_t)_height))
-                    PixelB = leds[XY(x, zF)];
-
-                leds[XY(x, y)] = (PixelA.nscale8(ease8InOutApprox(255 - fraction))) + (PixelB.nscale8(ease8InOutApprox(fraction)));
-            }
-        }
-    }
-#endif
-
-// This can't be in gfxbase.h because it uses the FillGetNoise() function template.
-void GFXBase::MoveInwardX(int startY, int endY)
-{
-    const int width = static_cast<int>(_width);
-    const int height = static_cast<int>(_height);
-    const int halfWidth = width / 2;
-    if (leds == nullptr || halfWidth <= 1 || height <= 0)
-        return;
-
-    startY = std::max(0, startY);
-    endY = std::min(height - 1, endY);
-
-    for (int y = startY; y <= endY; y++)
-    {
-        // Keep both halves inside the row. The previous loops read x+1 at
-        // width, which can corrupt unrelated pixels or memory at row edges.
-        for (int x = halfWidth - 1; x > 0; x--)
-            leds[XY(x, y)] = leds[XY(x - 1, y)];
-
-        for (int x = halfWidth; x < width - 1; x++)
-            leds[XY(x, y)] = leds[XY(x + 1, y)];
-    }
-}
-
-void GFXBase::MoveOutwardsX(int startY, int endY)
-{
-    const int width = static_cast<int>(_width);
-    const int height = static_cast<int>(_height);
-    const int halfWidth = width / 2;
-    if (leds == nullptr || halfWidth <= 1 || height <= 0)
-        return;
-
-    startY = std::max(0, startY);
-    endY = std::min(height - 1, endY);
-
-    for (int y = startY; y <= endY; y++)
-    {
-        for (int x = 0; x < halfWidth - 1; x++)
-            leds[XY(x, y)] = leds[XY(x + 1, y)];
-
-        for (int x = width - 1; x > halfWidth; x--)
-            leds[XY(x, y)] = leds[XY(x - 1, y)];
-    }
-}
-
-void GFXBase::MoveX(uint8_t delta) const
-{
-    for (int y = 0; y < (int)_height; y++)
-    {
-        for (int x = 0; x < (int)_width - delta; x++)
-            leds[XY(x, y)] = leds[XY(x + delta, y)];
-        for (int x = (int)_width - delta; x < (int)_width; x++)
-            leds[XY(x, y)] = leds[XY(x + delta - (int)_width, y)];
-    }
-}
-
-void GFXBase::MoveY(uint8_t delta) const
-{
-    CRGB tmp = 0;
-    for (int x = 0; x < (int)_width; x++)
-    {
-        tmp = leds[XY(x, 0)];
-        for (int m = 0; m < delta; m++)
-        {
-            for (int y = 0; y < (int)_height - 1; y++)
-                leds[XY(x, y)] = leds[XY(x, y + 1)];
-
-            leds[XY(x, (int)_height - 1)] = tmp;
-        }
-    }
-}
-
-void GFXBase::PrepareFrame()
-{
-}
-
-void GFXBase::PostProcessFrame(uint16_t, uint16_t)
-{
-}
-
 GFXBase::GFXBase(int w, int h) : Adafruit_GFX(w, h),
                         _width(w),
                         _height(h),
